+++ packages/web3-core/src/web3_config.ts
@@ -174,7 +174,8 @@ export abstract class Web3Config
 	}
 
 	/**
-	 * The time used to wait for Ethereum Node to return the sent transaction result. Note: If this method times out, the transaction may still be pending.
+	 * The time used to wait for Ethereum Node to return the sent transaction result.
+	 * Note: If the RPC call stuck at the Node and therefor timed-out, the transaction may still be pending or even mined by the Network. We recommend checking the pending transactions in such a case.
 	 * Default is `5000` ms.
 	 */
 	public get transactionRpcTimeout() {